pyproject.toml
[tool.poetry]
name = "oooscript"
version = "0.1.0"
description = "Description here"
authors = [":Barry-Thomas-Paul: Moss <vibrationoflife@protonmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
python-dotenv = "^0.21.0"
stickytape = "0.2.1"
pydantic = "^1.10.2"

[tool.poetry.group.dev.dependencies]
black = "^22.8.0"
typing-extensions = "^4.3.0"

[tool.poetry.group.test.dependencies]
pytest = "7.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
